points/datasets/auto_tagging/requirements.txt
pandas
onnxruntime
huggingface_hub
Pillow
opencv-contrib-python